Basic IO operations and Read from file example
Added examples on how to read an input from the keyboard using
Scanner class with System.in. With Scanner you have to use a different method
based on the type of input you are expecting.

You can also use the Scanner class with FileInputStream class to
read data from the user, which is also added in the example.

We will use Scanner temporarily for file operations, and will
switch to advanced versions.