Added Person Class
The Person class has an array of Dog, and an ArrayList of Cat. A Person
object can only adopt (adding to the array) 5 dogs at max using the
adoptDog(Dog dog) method. The Person class can also print the list of
dogs and cats they have. There isn't an add method for cats, because you
can use the ArrayList's built-in method to add/remove cats.